// frontend/src/app/api/models/usage-statistics-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { InvitationStatisticsResponse } from '../models/invitation-statistics-response';
import { ItemStatisticsResponse } from '../models/item-statistics-response';
import { StorageStatisticsResponse } from '../models/storage-statistics-response';
import { UserStatisticsResponse } from '../models/user-statistics-response';
export interface UsageStatisticsResponse {
  generatedAt: string;
  invitations: InvitationStatisticsResponse;
  items: ItemStatisticsResponse;
  storages: StorageStatisticsResponse;
  users: UserStatisticsResponse;
}